backend/requirements.txt
django
python-decouple
psycopg2-binary
dj-rest-auth
django-allauth
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
channels
requests
daphne
gunicorn
redis
django-redis
celery 
django-celery-beat 
django-celery-results
stripe
social-auth-core
reportlab
boto3
watchtower
channels-redis
drf-yasg
sentry-sdk[django]
pytest
pytest-mock